+++ src/config.cpp
@@ -26,7 +26,7 @@ int MAKE_PROGRESS = 0;
 int BIN_COUNT = 100;
 int BALANCE_LOAD = 0;
 //int BIN_CONTAINER = 0;
-float BALANCE_FACTOR = 1.5;
+double BALANCE_FACTOR = 1.5;
 
 // Profile & Debug
 int DBG_LEVEL = 0;
